Api/requirements.txt
blinker==1.6.2
certifi==2023.7.22
chardet==4.0.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
requests==2.25.1
urllib3==1.26.16
Werkzeug==2.3.7
